Do not publish if space has been manually edited
Any Confluence space that the plugin publishes specs to should be used
just for publishing specs, i.e. it should only contain specs and not
have any pages that were manually created by a user in Confluence.
This commit adds a safety check before publishing to ensure this. In a
future commit the plugin will delete all existing pages in the
Confluence space before publishing, so the safety check is an important
one (to ensure that no manually created pages are inadvertently
deleted).