Fix go linting by using latest version of linter
Prior to this commit the linter was skipping the linting check, but
still reporting success.

The linting was being skipped because the linter we were using an older
version of reviewdog's action-golangci-lint, which was not using the
latest version of Go (1.16) and therefore could not compile.

Updating to the latest version of action-golangci-lint causes the build
to fail as it should when there is an error like this.

Also specified the Go version used by the linter to be 1.16 explicitly.