Add logging capability to the plugin
Debug and info are the only logging levels provided, as per this advice:
https://dave.cheney.net/2015/11/05/lets-talk-about-logging

The log level is set by setting `GAUGE_LOG_LEVEL` (as an environment
variable or a property) to `debug` or `info` (default is`info`).
It controls the logging level both for the log files which are
generated, _and_ what is logged to the console.

NB the command line flag
`--log-level` does not have any effect on the logging for this plugin,
it just effects the log level on core Gauge.

The logging implementation is lifted and shifted with just some minor
tweaks [from core Gauge's implementation][1] (couldn't import it
directly as a module as it would clash with the gauge-proto git
submodule).

[1]: https://github.com/getgauge/gauge/tree/master/logger