Do not publish if space has been manually edited (#16)
* Do not publish if space has been manually edited

Any Confluence space that the plugin publishes specs to should be used
just for publishing specs, i.e. it should only contain specs and not
have any pages that were manually created by a user in Confluence.
This commit adds a safety check before publishing to ensure this. In a
future commit the plugin will delete all existing pages in the
Confluence space before publishing, so the safety check is an important
one (to ensure that no manually created pages are inadvertently
deleted).

* Refactor by extracting homepage type

* Refactor publisher and space types

* Refactor space setup

* Refactor lastPublished code

* Refactor cqloffset

* Add debug logging

* Add "retry" to allowed commit message first words

* Retry cql time offset calculations up to 5 times

* Document the CQL offset functionality

* Bump plugin minor version

* Fix typo in spec