Ensure test Confluence spaces have unique keys
Prior to this commit we were using the current time in milliseconds as
the Confluence space key in our functional tests.  There was a chance
though that there could be naming clashes when running multiple tests in
parallel, so this commit uses a [UUID][1] as the unique identifier
instead.

This change is just in our functional tests code, not in the actual
plugin production code.

[1]: https://www.baeldung.com/java-uuid