Fix CQL time calculation bug caused by pagination
This commit fixes an intermittent bug which appeared when there were
sufficient specs already published in the Confluence space to cause
pagination in the CQL search used to calculate the CQL time.  The fix
is to constrain the search by page ID to only return the homepage, not
all the pages that were created at the given time.  That means that the
results of the search will never be paginated.