Delete existing published specs before publishing (#22)
* Delete existing published specs before publishing

It is safe to do this as before doing so we abort if the Space has been
manually edited since the last publish.

NB the Space homepage is not deleted, just the published specs (i.e. all
other pages except the homepage).

* Fix CQL time calculation bug caused by pagination

This commit fixes an intermittent bug which appeared when there were
sufficient specs already published in the Confluence space to cause
pagination in the CQL search used to calculate the CQL time.  The fix
is to constrain the search by page ID to only return the homepage, not
all the pages that were created at the given time.  That means that the
results of the search will never be paginated.

* Bump plugin minor version