Fix bug when republishing after aborted attempt
The bug was caused because prior to this commit we were only
incrementing the lastPublished counter if publishing succeeded. This
meant that for instance if publishing was aborted the very first time
that the plugin was ran, it would fail again afterwards as it would
mistakenly think there had been a manual edit of the Confluence space.

The fix in this commit is to ensure that we also increment the
lastPublished counter when there is an error encountered during
publishing.