Remove debugging output from console (#34)
* Remove debugging output from console

Prior to this commit we were displaying a number of debugging messages
as console output when the plugin was invoked.  This was not ideal as
the messages clutter the console output and would not be helpful to
standard users of the plugin. Users of the plugin no longer see these
messages on the console now.  The messages still appear in the
`gaugeconfluence.log` though, which is what we want.

Fixes #25

* Bump plugin patch version