Fix error message when Space homepage is missing
This commit fixes a bug where a misleading error message was being
displayed when the Confluence Space has no homepage.  The error now
asks the user of the plugin to add a homepage manually to the Confluence
Space and to try again.

In a future merge request we aim to improve how the plugin handles this
scenario, by automatically adding a homepage instead of reporting an
error.  As we like [small atomic commits][1] this commit is the first
step in the right direction here.

[1]: https://tekin.co.uk/2021/01/how-atomic-commits-make-you-a-better-coder