Move space setup to after pre-publish checks
This means that the dry-run mode is guaranteed not to have side-effects
as there is no interaction at all with the Confluence Space during the
pre-publish checks.  It also means that the Confluence Space does not
need to exist when running in dry-run mode, which is useful to help
people get started in trying out the plugin.