Disable tests on Confluence Server temporarily (#56)
* Disable tests on Confluence Server temporarily

Up until now we have always run the Continuous Integration tests
against both Confluence Cloud and Confluence Server. There is a
temporary issue with the instance of Confluence Server though (it runs
on Google Cloud Platform and has been mistakenly flagged by Google as
doing Cryptocurrency mining for some reason).  We will reinstate the
tests on Confluence Server as soon as we have a test instance available
again.

There have not been any discrepancies at all between the behaviour on
Confluence Cloud and Confluence Server, which bodes well in terms of
hopefully not finding any errors when we do reinstate Confluence Server.

There is one test to do with Space permissions which we can only run on
Server and not Cloud, so that test will not be run again until we have
Server tests reinstated.

* Do not run Gauge specs in parallel

Running the specs in parallel may have been the cause of intermittent
test failures.

* Handle CQL time calculation error properly

Prior to this commit we were silently ignoring any error during the CQL
calculation time, which was causing a difficult to diagnose panic error
if the CQL calculation code failed.

* Bump plugin patch version