Remove the need to specify a Confluence Space key (#57)
* Remove the need to specify a Confluence Space key

Prior to this commit users of the plugin always had to specify the key
of the Confluence Space that they wanted to be published to (and to be
created if necessary).  This commit makes the `CONFLUENCE_SPACE_KEY`
Gauge configuration variable optional rather than mandatory.

If the `CONFLUENCE_SPACE_KEY` is not provided, the plugin will derive
the Space key to be used based on the remote Git repository URL. This
convention ensures that each Git repository has its own unique
Confluence space key derived from it, i.e. a one to one mapping between
each Git repository and its associated one to one space.

The recommended way to run the plugin now is not to provide the
`CONFLUENCE_SPACE_KEY` variable, and instead to rely on the plugin to
set it.  This is particularly useful in CI/CD for instance, as it
removes the need to set the Space key manually before being able to run
the plugin.

One use case for setting the `CONFLUENCE_SPACE_KEY` is if for whatever
reason you are unable to specify a Confluence user who has permission
to create Confluence Spaces.  By setting the `CONFLUENCE_SPACE_KEY` to
be an existing Space which someone (e.g. a Confluence admin) has
created for you, you will still be able to run the plugin even without
Confluence create space permissions.

* Refactor tests so git remote url can be specified

* Bump plugin minor version