Improve Space homepage content
Prior to this commit the Confluence Space homepage (for any Spaces that
we publish Gauge specs to) was just Confluence's default generic
homepage.  This commit updates the Confluence Space homepage to have a
useful overview of what the Space contains and why, along with links to
find out more about Gauge and this Gauge Confluence plugin.

We republish this homepage content on every run of the plugin, even
though it is just rewriting the same content each time.  This is so that
it is easy to update the homepage content in the future, if and when we
want to amend or improve it.