Test that homepage is republished on every run
This commit adds a Gauge spec to verify that the Confluence Space
homepage is updated (i.e. republished by the plugin) on every run of the
plugin.

This is important as if we tried to do something "clever" such as only
updating the homepage on the first run of the plugin, we'd be stuck if
we ever want to roll out a change which modifies the content of the
homepage.