Run the CI tests just on Ubuntu
Prior to this commit we were running the CI tests on Ubuntu, Windows
and macOS.  This commit removes Windows and macOS from the CI/CD tests.
The reason for this is that having the test suite run simultaneously on
all three was sometimes causing subtle test failures, generally caused
by contention around the same Confluence Space.

Also created #62 to ensure that we run the CI tests on a schedule
(weekly?) against all 3 operating systems.