Improve Space homepage content (#61)
* Improve Space homepage content

Prior to this commit the Confluence Space homepage (for any Spaces that
we publish Gauge specs to) was just Confluence's default generic
homepage.  This commit updates the Confluence Space homepage to have a
useful overview of what the Space contains and why, along with links to
find out more about Gauge and this Gauge Confluence plugin.

We republish this homepage content on every run of the plugin, even
though it is just rewriting the same content each time.  This is so that
it is easy to update the homepage content in the future, if and when we
want to amend or improve it.

* Fix functional tests by updating Space name

* Set default Confluence Space key in tests

* Test that homepage is republished on every run

This commit adds a Gauge spec to verify that the Confluence Space
homepage is updated (i.e. republished by the plugin) on every run of the
plugin.

This is important as if we tried to do something "clever" such as only
updating the homepage on the first run of the plugin, we'd be stuck if
we ever want to roll out a change which modifies the content of the
homepage.

* Bump plugin minor version

* Run the CI tests just on Ubuntu

Prior to this commit we were running the CI tests on Ubuntu, Windows
and macOS.  This commit removes Windows and macOS from the CI/CD tests.
The reason for this is that having the test suite run simultaneously on
all three was sometimes causing subtle test failures, generally caused
by contention around the same Confluence Space.

Also created #62 to ensure that we run the CI tests on a schedule
(weekly?) against all 3 operating systems.