Ensure plugin Git branch logic works on GitLab CI (#64)
* Ensure plugin Git branch logic works on GitLab CI

This commit fixes an issue when running the plugin on GitLab CI. Prior
to this commit the plugin would report that it could not obtain the
current Git branch when running on GitLab CI. This was because GitLab CI
runs in detached HEAD mode, meaning that the plugin's Git branch
derivation logic could not work.

(The plugin needs to derive the Git branch so that it can provide a link
from each published Gauge spec in Confluence back to the spec file's
source code in GitLab/GitHub etc.)

The fix is to look for a CI_COMMIT_BRANCH environment variable (which
is automatically set by GitLab CI), and if it exists then to use it as
the current branch name.  If the environment variable does not exist
(i.e. when not running on GitLab) then the logic falls back to deriving
the Git branch from the local Git files as it always has done.

* Bump plugin patch version