Fix Confluence Cloud line breaks bug
Prior to this commit the published pages in Confluence Cloud contained
extra line breaks which should not have been there.  Line breaks were
being added at every line break in the markdown, rather than being
interpreted as spaces as they should be in markdown.

This fix does not alter the behaviour at all on Confluence Server, as
the bug was only occurring on Confluence Cloud.